Moved pages outside of Component.
To have the URL path no longer be component/, moved the pages out. Also edited the href to also go to correct new URL.
