+++ src/app/BookAppointment/page.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "../Navbar";
+import Navbar from "../components/Navbar";
 
 const Booking = () => {
   return <Navbar />;
