+++ src/app/page.js
@@ -1,5 +1,5 @@
 import Image from 'next/image'
-import HomeView from './components/Home/HomeView'
+import HomeView from '@/app/Home/HomeView'
 
 export default function Home() {
   return (
